Only read local_config once
Without this change, filter_poodll is performing lookups for config far
more than necessary. This change reduces it to *one* read total per page
view.

On any given page, it should only need to load the course config *once*.
However without this patch we actually see:

course with no set poodll config:
* 8*invocations
course with set poodll config:
* 1*invocations

Where an "invocation" is a call to the filter - so for every label and
such on the page. We're seeing an average of 50 DB reads for the same
data in course views as a result.